requirements.txt
asgiref==3.7.2
asttokens==2.4.1
attrs==24.2.0
beautifulsoup4==4.12.3
Brotli==1.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.17.1
charset-normalizer==3.3.2
comm==0.2.2
cssselect2==0.8.0
debugpy==1.8.1
decorator==5.1.1
Django==5.0
django-cors-headers==4.6.0
django-storages==1.14.2
djangorestframework==3.15.1
drf-spectacular==0.27.2
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
fonttools==4.57.0
google-api-core==2.15.0
google-auth==2.25.2
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==2.11.0
google-crc32c==1.7.1
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
greenlet==3.0.3
gspread==6.1.2
html5lib==1.1
httplib2==0.22.0
idna==3.6
inflection==0.5.1
ipykernel==6.29.4
ipython==8.25.0
jedi==0.19.1
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mysql-connector-python==8.1.0
mysqlclient==2.2.7,<2.3
nest-asyncio==1.6.0
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
openpyxl==3.1.2
packaging==24.1
pandas==2.2.2
parso==0.8.4
pdfkit==1.0.0
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==4.21.12
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydyf==0.11.0
Pygments==2.18.0
pyparsing==3.1.4
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-magic==0.4.27
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.21.0
rsa==4.9
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.19
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.9.0
tzdata==2024.1
Unidecode==1.3.8
uritemplate==4.1.1
urllib3==2.1.0
validate-docbr==1.10.0
wcwidth==0.2.13
weasyprint==53.0
webencodings==0.5.1
XlsxWriter==3.2.0
zopfli==0.2.3.post1
